Add: #231 Wave 1 세션 D — feature/home/impl Android-only 추출 (billing/ads/credential platform 이전)
- platform/billing/{api,impl}: BillingService interface + AndroidBillingService (BillingManager 로직 이전)
- platform/ads/{api,impl}: AdsService interface + AndroidAdsService (RewardAdManager 로직 이전, MobileAds.initialize 포함)
- platform/credential/{api,impl}: CredentialService + AndroidCredentialService (Google Sign-In)
- feature/home/impl/data/{ads,billing} 제거, Android 라이브러리 의존 제거 후 platform-impl 참조 추가
- 콜사이트 Koin inject로 교체 (MainActivity/MemozyApplication/LoginScreen/SettingsScreen/DonationScreen/SubscriptionScreen)
- feature/core/resource: LocalRewardAdProvider 타입을 AdsService?로 교체, google-fonts api→implementation

Runtime smoke: 리뷰어 로컬 확인 필요 (홈 리워드 광고 / Google 로그인 / 구독 / 후원).
