+++ feature/home/impl/build.gradle.kts
@@ -34,18 +34,16 @@ dependencies {
     implementation(projects.datasource.remote.auth.api)
     implementation(projects.datasource.remote.ai.api)
     implementation(projects.data.backup.api)
+    implementation(projects.platform.billing.api)
+    implementation(projects.platform.ads.api)
+    implementation(projects.platform.credential.api)
     implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.activity.compose)
     implementation(libs.androidx.compose.navigation)
     implementation(libs.haze)
     implementation(libs.kotlinx.datetime)
     implementation(libs.montage.android)
-    implementation(libs.billing.ktx)
-    implementation(libs.play.services.ads)
     implementation(libs.kotlinx.coroutines.android)
     implementation(libs.androidx.compose.material.icons.extended)
     implementation(libs.koin.androidx.compose)
-    implementation(libs.credential.manager)
-    implementation(libs.credential.manager.play)
-    implementation(libs.google.id.identity)
 }
