+++ feature/home/impl/src/main/java/me/pecos/memozy/presentation/screen/settings/SettingsScreen.kt
@@ -37,11 +37,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
-import androidx.credentials.CredentialManager
-import androidx.credentials.GetCredentialRequest
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.google.android.libraries.identity.googleid.GetGoogleIdOption
-import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
 import androidx.compose.foundation.background
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.width
@@ -82,7 +78,10 @@ import me.pecos.memozy.feature.core.viewmodel.settings.CloudBackupState
 import me.pecos.memozy.feature.core.viewmodel.settings.FontSizeLevel
 import me.pecos.memozy.feature.core.viewmodel.settings.LANGUAGES
 import me.pecos.memozy.feature.core.viewmodel.settings.ThemeMode
+import me.pecos.memozy.platform.credential.CredentialService
+import me.pecos.memozy.platform.credential.GoogleSignInResult
 import me.pecos.memozy.presentation.theme.LocalActivity
+import org.koin.compose.koinInject
 import me.pecos.memozy.presentation.theme.LocalAppColors
 import me.pecos.memozy.presentation.theme.LocalSubscriptionTier
 import me.pecos.memozy.presentation.theme.LocalFontSettings
@@ -119,6 +118,7 @@ fun SettingsScreen(
     val context = LocalContext.current
     val activity = LocalActivity.current
     val scope = rememberCoroutineScope()
+    val credentialService: CredentialService = koinInject()
 
     // 로그인 상태 변경 시 마지막 백업 시간 로드
     LaunchedEffect(authState) {
@@ -603,32 +603,25 @@ fun SettingsScreen(
                             colors = ButtonDefaults.outlinedButtonColors(contentColor = colors.textTitle),
                             onClick = {
                                 scope.launch {
-                                    try {
-                                        val credentialManager = CredentialManager.create(context)
-                                        val googleIdOption = GetGoogleIdOption.Builder()
-                                            .setServerClientId(BuildConfig.GOOGLE_WEB_CLIENT_ID)
-                                            .setFilterByAuthorizedAccounts(false)
-                                            .build()
-                                        val request = GetCredentialRequest.Builder()
-                                            .addCredentialOption(googleIdOption)
-                                            .build()
-                                        val result = credentialManager.getCredential(
-                                            context = activity ?: context,
-                                            request = request,
-                                        )
-                                        val googleIdToken = GoogleIdTokenCredential
-                                            .createFrom(result.credential.data)
-                                            .idToken
-                                        settingsViewModel.signInWithGoogle(googleIdToken)
-                                    } catch (e: androidx.credentials.exceptions.GetCredentialCancellationException) {
-                                        // 사용자가 취소 — 무시
-                                    } catch (e: Exception) {
-                                        android.util.Log.e("SettingsAuth", "Sign-in failed", e)
-                                        Toast.makeText(
-                                            context,
-                                            context.getString(R.string.sign_in_error),
-                                            Toast.LENGTH_SHORT
-                                        ).show()
+                                    val result = credentialService.signInWithGoogle(
+                                        activity = activity ?: context,
+                                        serverClientId = BuildConfig.GOOGLE_WEB_CLIENT_ID,
+                                    )
+                                    when (result) {
+                                        is GoogleSignInResult.Success ->
+                                            settingsViewModel.signInWithGoogle(result.idToken)
+                                        is GoogleSignInResult.Cancelled -> Unit
+                                        is GoogleSignInResult.Error -> {
+                                            android.util.Log.e(
+                                                "SettingsAuth",
+                                                "Sign-in failed: ${result.message}"
+                                            )
+                                            Toast.makeText(
+                                                context,
+                                                context.getString(R.string.sign_in_error),
+                                                Toast.LENGTH_SHORT
+                                            ).show()
+                                        }
                                     }
                                 }
                             }
