+++ platform/ads/api/src/commonMain/kotlin/me/pecos/memozy/platform/ads/AdsService.kt
@@ -1,3 +1,13 @@
 package me.pecos.memozy.platform.ads
 
-interface AdsService
+interface AdsService {
+    val isAdReady: Boolean
+    val isAdLoading: Boolean
+
+    fun initialize()
+    fun loadAd()
+    fun showAd(onRewarded: () -> Unit)
+
+    fun bindActivity(activity: Any?)
+    fun unbindActivity()
+}
