+++ platform/credential/api/src/commonMain/kotlin/me/pecos/memozy/platform/credential/CredentialService.kt
@@ -1,3 +1,14 @@
 package me.pecos.memozy.platform.credential
 
-interface CredentialService
+sealed class GoogleSignInResult {
+    data class Success(val idToken: String) : GoogleSignInResult()
+    data object Cancelled : GoogleSignInResult()
+    data class Error(val message: String) : GoogleSignInResult()
+}
+
+interface CredentialService {
+    suspend fun signInWithGoogle(
+        activity: Any?,
+        serverClientId: String,
+    ): GoogleSignInResult
+}
